refactor: migrate useUserStore and userProvider from c-store to store directory
- Moved useUserStore.ts and userProvider.tsx from src/c-store/ to src/store/
- Updated all 22 import statements across the codebase to use the new @/store path
- Added exports to store/index.ts for the migrated modules
- Removed obsolete exports from c-store/index.ts
- Ensures consistency with the project's directory naming conventions

This refactoring aligns with the standardized directory structure and helps consolidate store-related modules under a unified location.
