+++ src/cook-web/src/app/layout.tsx
@@ -5,7 +5,7 @@ import './globals.css';
 import '@/assets/css/md-editor.css';
 import '@/assets/css/markdown.css';
 import { ConfigProvider } from '@/components/config-provider';
-import { UserProvider } from '@/c-store/userProvider';
+import { UserProvider } from '@/store/userProvider';
 import '@/i18n';
 
 const geistSans = Geist({
