+++ src/cook-web/src/c-store/index.ts
@@ -3,5 +3,3 @@ export * from './envStore';
 export * from './useCourseStore';
 export * from './useSystemStore';
 export * from './useUiLayoutStore';
-export * from './useUserStore';
-export * from './userProvider';
