+++ src/cook-web/src/components/main-root/MainRoot.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { UserProvider } from '@/c-store';
+import { UserProvider } from '@/store/userProvider';
 import React from 'react';
 import ShifuEdit from '../shifu-edit';
